[Update] reflection and refraction shader

+++ src/tools/ModelLoader/Mesh.cpp
@@ -48,7 +48,7 @@ void Mesh::Draw(Shader& shader)
         else if (name == "texture_specular")
             number = std::to_string(specularNr++);
 
-        shader.SetFloat((name + number).c_str(), i);
+        shader.SetInt((name + number).c_str(), i);
        
 
         textures[i]->Bind(i);
